# fix `ptx` linearizer bug 2 [pr]