@ignaciosica 2025-04-21
Keeping as a draft until I add regression test. While adding TC=3 job in benchmark I realized the use_tensor_core attribute from apply_tensor_cores was not propagating correctly. Working on fixing.